+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Development
 * Report when file directory does not exist.
 * Better testing for prohibited file types.
 * Only create database after file path verified.
+* Add ``force_ocr`` parameter.
 
 0.1.1 - December 31, 2015
 -------------------------